python modules match typescript modules
